pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=45.0.2",
    "flask>=3.1.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "mysql-connector-python>=9.3.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
]
